back-end/pyproject.toml
[project]
name = "back-end"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "hnswlib>=0.8.0",
    "icecream>=2.1.4",
    "itsdangerous>=2.2.0",
    "librosa>=0.11.0",
    "pandas>=2.2.3",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.3",
    "pydub>=0.25.1",
    "pyjwt>=2.10.1",
    "pysoundfile>=0.9.0.post1",
    "python-dotenv>=1.1.0",
    "pytube>=15.0.0",
    "pytubefix>=8.13.1",
    "sqladmin>=0.20.1",
    "sqlalchemy[asyncio]>=2.0.40",
    "uuid6>=2024.7.10",
    "uvicorn[all]>=0.34.2",
]
